Switch to using ghcr.io registry to store images instead of Dockerhub
For now, storing metadata tags in artifact file since now we can avoid using timestamp and use tags / release versions.
But would still need to store these versions so we can pull relevant images.

Or, an alternative would be to use the latest tagged image which would have both a version and latest tag.
But if only latest mentioned in image to pull then we would be using unversioned images and might not be straightforward to identify which image was "latest" in the past.

So might have to stick to using versioned tags, releases, or SHA image digests.